Reorganize client files
