+++ client/src/styles/components/_Avatar.scss
