+++ client/src/styles/components/_ChooseAvatar.scss
