+++ client/src/styles/components/_quiz.scss
