--- client/src/styles/layouts/App.scss
@@ -1,10 +1,0 @@
-/**
- * Main contain the content of each page
- */
-main {
-  padding: 0;
-  margin: 0;
-  min-height: calc(100% - 50px);
-  background-color: $bg;
-  box-sizing: border-box;
-}
