+++ client/src/styles/styles.scss
@@ -1,8 +1,7 @@
-@import "base/variables", "base/reset", "base/main", "base/shared", "base/mixins";
+// Keep this order for these base imports
+@import "base/reset", "base/variables", "base/mixins", "base/main", "base/shared";
 
-@import "components/TopBar", "components/buttons", "components/messages", "components/userBadge", "components/avatar",
-  "components/chooseAvatar";
-
-@import "layouts/App", "layouts/quiz";
+@import "components/Avatar", "components/buttons", "components/ChooseAvatar", "components/quiz", "components/TopBar",
+  "components/userBadge";
 
 @import "pages/Menu", "pages/About", "pages/Profile", "pages/Login";
