+++ client/src/test/App.test.js
@@ -2,6 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { expect } from "chai";
 import { render } from "@testing-library/react";
+
 import App from "../App";
 
 /* eslint-disable */
