+++ client/src/test/Login.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import { shallow } from "enzyme";
 import { expect } from "chai";
 
-import Login from "../components/pages/Login";
+import Login from "../pages/Login";
 
 describe("Well displayed Login component", function () {
   it("Should contains 2 text inputs and 1 button", function (done) {
